chore: keep function private and move unit test into new file
