+++ go/adbc/sqldriver/driver_internals_test.go
@@ -1,0 +1,44 @@
+package sqldriver
+
+import (
+	"fmt"
+	"strings"
+	"testing"
+
+	"github.com/apache/arrow-adbc/go/adbc"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestParseConnectStr(t *testing.T) {
+	const (
+		scheme   = "grpc+tcp"
+		host     = "host"
+		port     = 443
+		dbname   = "dbname"
+		username = "username"
+		password = "token=="
+	)
+
+	var (
+		uri = fmt.Sprintf("%s://%s:%d/%s", scheme, host, port, dbname)
+	)
+
+	dsn := strings.Join([]string{
+		fmt.Sprintf("%s=%s", adbc.OptionKeyURI, uri),
+		fmt.Sprintf("%s=%s", adbc.OptionKeyUsername, username),
+		fmt.Sprintf("%s=%s", adbc.OptionKeyPassword, password),
+		fmt.Sprintf("%s=%s", adbc.OptionKeyReadOnly, adbc.OptionValueEnabled),
+	}, " ; ")
+
+	expectOpts := map[string]string{
+		adbc.OptionKeyURI:      uri,
+		adbc.OptionKeyUsername: username,
+		adbc.OptionKeyPassword: password,
+		adbc.OptionKeyReadOnly: adbc.OptionValueEnabled,
+	}
+
+	gotOpts, err := parseConnectStr(dsn)
+	if assert.NoError(t, err) {
+		assert.Equal(t, expectOpts, gotOpts)
+	}
+}
