+++ go/adbc/sqldriver/driver_test.go
@@ -22,10 +22,7 @@ package sqldriver_test
 import (
 	"database/sql"
 	"fmt"
-	"strings"
-	"testing"
 
-	"github.com/apache/arrow-adbc/go/adbc"
 	"github.com/apache/arrow-adbc/go/adbc/drivermgr"
 	"github.com/apache/arrow-adbc/go/adbc/sqldriver"
 )
@@ -73,38 +70,4 @@ func Example() {
 	// 1
 	// true true
 	// 1
-}
-
-func TestParseConnectStr(t *testing.T) {
-	const (
-		scheme   = "grpc+tcp"
-		host     = "host"
-		port     = 443
-		dbname   = "dbname"
-		username = "username"
-		password = "token=="
-	)
-
-	var (
-		uri = fmt.Sprintf("%s://%s:%d/%s", scheme, host, port, dbname)
-	)
-
-	dsn := strings.Join([]string{
-		fmt.Sprintf("%s=%s", adbc.OptionKeyURI, uri),
-		fmt.Sprintf("%s=%s", adbc.OptionKeyUsername, username),
-		fmt.Sprintf("%s=%s", adbc.OptionKeyPassword, password),
-		fmt.Sprintf("%s=%s", adbc.OptionKeyReadOnly, adbc.OptionValueEnabled),
-	}, " ; ")
-
-	expectOpts := map[string]string{
-		adbc.OptionKeyURI:      uri,
-		adbc.OptionKeyUsername: username,
-		adbc.OptionKeyPassword: password,
-		adbc.OptionKeyReadOnly: adbc.OptionValueEnabled,
-	}
-
-	gotOpts, err := sqldriver.ParseConnectStr(dsn)
-	if assert.NoError(t, err) {
-		assert.Equal(t, expectOpts, gotOpts)
-	}
 }
